+++ src/pseudopeople/interface.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from typing import Union
+from typing import Dict, Union
 
 import pandas as pd
 import pyarrow.parquet as pq
@@ -8,7 +8,8 @@ from loguru import logger
 from pseudopeople.configuration import get_configuration
 from pseudopeople.constants import paths
 from pseudopeople.noise import noise_form
-from pseudopeople.schema_entities import FORMS, Form
+from pseudopeople.schema_entities import COLUMNS, FORMS, Form
+from pseudopeople.utilities import configure_logging_to_terminal
 
 
 def _generate_form(
@@ -16,7 +17,8 @@ def _generate_form(
     source: Union[Path, str],
     seed: int,
     configuration: Union[Path, str, dict],
-    year_filter: dict,
+    year_filter: Dict,
+    verbose: bool = False,
 ) -> pd.DataFrame:
     """
     Helper for generating noised forms from clean data.
@@ -29,9 +31,14 @@ def _generate_form(
         Seed for controlling randomness
     :param configuration:
         Object to configure noise levels
+    :param year_filter:
+        Dictionary with keys 'hdf' and 'parquet' and values filter lists
+    :param verbose:
+        Log with verbosity if True. Default is False.
     :return:
         Noised form data in a pd.DataFrame
     """
+    configure_logging_to_terminal(verbose)
     configuration_tree = get_configuration(configuration)
     # TODO: we should save outputs of the simulation with filenames that are
     #  consistent with the names of the forms if possible.
@@ -46,11 +53,10 @@ def _generate_form(
     source = Path(source) / form_file_name
     data_paths = [x for x in source.glob(f"{form_file_name}*")]
     if not data_paths:
-        logger.warning(
+        raise ValueError(
             f"No datasets found at directory {str(source)}. "
             "Please provide the path to the unmodified root data directory."
         )
-        return None
     suffix = set(x.suffix for x in data_paths)
     if len(suffix) > 1:
         raise TypeError(
@@ -58,43 +64,65 @@ def _generate_form(
             "Please provide the path to the unmodified root data directory."
         )
     noised_form = []
-    columns_to_keep = [c for c in form.columns]
     for data_path in data_paths:
-        if data_path.suffix == ".hdf":
-            with pd.HDFStore(str(data_path), mode="r") as hdf_store:
-                data = hdf_store.select("data", where=year_filter["hdf"])
-        elif data_path.suffix == ".parquet":
-            data = pq.read_table(data_path, filters=year_filter["parquet"]).to_pandas()
-        else:
-            raise ValueError(
-                "Source path must either be a .hdf or a .parquet file. Provided "
-                f"{data_path.suffix}"
-            )
-        if not isinstance(data, pd.DataFrame):
-            raise TypeError(
-                f"File located at {data_path} must contain a pandas DataFrame. "
-                "Please provide the path to the unmodified root data directory."
-            )
+        logger.info(f"Loading data from {data_path}.")
+        data = _load_data_from_path(data_path, year_filter)
 
-        # Coerce dtypes prior to noising to catch issues early as well as
-        # get most columns away from dtype 'category' and into 'object' (strings)
-        for col in columns_to_keep:
-            if col.dtype_name != data[col.name].dtype.name:
-                data[col.name] = data[col.name].astype(col.dtype_name)
-
+        data = _reformat_dates_for_noising(data, form)
+        data = _coerce_dtypes(data, form)
         noised_data = noise_form(form, data, configuration_tree, seed)
-        noised_data = _extract_columns(columns_to_keep, noised_data)
+        noised_data = _extract_columns(form.columns, noised_data)
         noised_form.append(noised_data)
 
     noised_form = pd.concat(noised_form, ignore_index=True)
 
     # Known pandas bug: pd.concat does not preserve category dtypes so we coerce
     # again after concat (https://github.com/pandas-dev/pandas/issues/51362)
-    for col in columns_to_keep:
-        if col.dtype_name != noised_form[col.name].dtype.name:
-            noised_form[col.name] = noised_form[col.name].astype(col.dtype_name)
+    noised_form = _coerce_dtypes(noised_form, form)
 
     return noised_form
+
+
+def _coerce_dtypes(data: pd.DataFrame, form: Form):
+    # Coerce dtypes prior to noising to catch issues early as well as
+    # get most columns away from dtype 'category' and into 'object' (strings)
+    for col in form.columns:
+        if col.dtype_name != data[col.name].dtype.name:
+            data[col.name] = data[col.name].astype(col.dtype_name)
+    return data
+
+
+def _load_data_from_path(data_path: Path, year_filter: Dict):
+    """Load data from a data file given a data_path and a year_filter."""
+    if data_path.suffix == ".hdf":
+        with pd.HDFStore(str(data_path), mode="r") as hdf_store:
+            data = hdf_store.select("data", where=year_filter["hdf"])
+    elif data_path.suffix == ".parquet":
+        data = pq.read_table(data_path, filters=year_filter["parquet"]).to_pandas()
+    else:
+        raise ValueError(
+            "Source path must either be a .hdf or a .parquet file. Provided "
+            f"{data_path.suffix}"
+        )
+    if not isinstance(data, pd.DataFrame):
+        raise TypeError(
+            f"File located at {data_path} must contain a pandas DataFrame. "
+            "Please provide the path to the unmodified root data directory."
+        )
+    return data
+
+
+def _reformat_dates_for_noising(data: pd.DataFrame, form: Form):
+    """Formats SSA event_date and dates of birth, so they can be noised."""
+    if COLUMNS.ssa_event_date.name in data.columns and form == FORMS.ssa:
+        # event_date -> YYYYMMDD
+        data[COLUMNS.ssa_event_date.name] = data[COLUMNS.ssa_event_date.name].dt.strftime(
+            "%Y%m%d"
+        )
+    if COLUMNS.dob.name in data.columns:
+        # date_of_birth -> MM/DD/YYYY
+        data[COLUMNS.dob.name] = data[COLUMNS.dob.name].dt.strftime("%m/%d/%Y")
+    return data
 
 
 def _extract_columns(columns_to_keep, noised_form):
@@ -104,12 +132,12 @@ def _extract_columns(columns_to_keep, noised_form):
     return noised_form
 
 
-# TODO: add year as parameter to select the year of the decennial census to generate (MIC-3909)
 def generate_decennial_census(
     source: Union[Path, str] = None,
     seed: int = 0,
     configuration: Union[Path, str, dict] = None,
     year: int = 2020,
+    verbose: bool = False,
 ) -> pd.DataFrame:
     """
     Generates noised decennial census data from un-noised data.
@@ -118,13 +146,14 @@ def generate_decennial_census(
     :param seed: An integer seed for randomness
     :param configuration: (optional) A path to a configuration YAML file or a dictionary to override the default configuration
     :param year: The year from the data to noise
+    :param verbose: Log with verbosity if True. Default is False.
     :return: A pd.DataFrame of noised census data
     """
     year_filter = {"hdf": None, "parquet": None}
     if year:
         year_filter["hdf"] = [f"{FORMS.census.date_column} == {year}."]
         year_filter["parquet"] = [(FORMS.census.date_column, "==", year)]
-    return _generate_form(FORMS.census, source, seed, configuration, year_filter)
+    return _generate_form(FORMS.census, source, seed, configuration, year_filter, verbose)
 
 
 def generate_american_communities_survey(
@@ -132,6 +161,7 @@ def generate_american_communities_survey(
     seed: int = 0,
     configuration: Union[Path, str, dict] = None,
     year: int = 2020,
+    verbose: bool = False,
 ) -> pd.DataFrame:
     """
     Generates noised American Communities Survey (ACS) data from un-noised data.
@@ -140,6 +170,7 @@ def generate_american_communities_survey(
     :param seed: An integer seed for randomness
     :param configuration: (optional) A path to a configuration YAML file or a dictionary to override the default configuration
     :param year: The year from the data to noise
+    :param verbose: Log with verbosity if True. Default is False.
     :return: A pd.DataFrame of noised ACS data
     """
     year_filter = {"hdf": None, "parquet": None}
@@ -152,7 +183,7 @@ def generate_american_communities_survey(
             (FORMS.acs.date_column, "<=", pd.Timestamp(f"{year}-12-31")),
         ]
         seed = seed * 10_000 + year
-    return _generate_form(FORMS.acs, source, seed, configuration, year_filter)
+    return _generate_form(FORMS.acs, source, seed, configuration, year_filter, verbose)
 
 
 def generate_current_population_survey(
@@ -160,6 +191,7 @@ def generate_current_population_survey(
     seed: int = 0,
     configuration: Union[Path, str, dict] = None,
     year: int = 2020,
+    verbose: bool = False,
 ) -> pd.DataFrame:
     """
     Generates noised Current Population Survey (CPS) data from un-noised data.
@@ -168,6 +200,7 @@ def generate_current_population_survey(
     :param seed: An integer seed for randomness
     :param configuration: (optional) A path to a configuration YAML file or a dictionary to override the default configuration
     :param year: The year from the data to noise
+    :param verbose: Log with verbosity if True. Default is False.
     :return: A pd.DataFrame of noised CPS data
     """
     year_filter = {"hdf": None, "parquet": None}
@@ -180,7 +213,7 @@ def generate_current_population_survey(
             (FORMS.cps.date_column, "<=", pd.Timestamp(f"{year}-12-31")),
         ]
         seed = seed * 10_000 + year
-    return _generate_form(FORMS.cps, source, seed, configuration, year_filter)
+    return _generate_form(FORMS.cps, source, seed, configuration, year_filter, verbose)
 
 
 def generate_taxes_w2_and_1099(
@@ -188,6 +221,7 @@ def generate_taxes_w2_and_1099(
     seed: int = 0,
     configuration: Union[Path, str, dict] = None,
     year: int = 2020,
+    verbose: bool = False,
 ) -> pd.DataFrame:
     """
     Generates noised W2 and 1099 data from un-noised data.
@@ -196,6 +230,7 @@ def generate_taxes_w2_and_1099(
     :param seed: An integer seed for randomness
     :param configuration: (optional) A path to a configuration YAML file or a dictionary to override the default configuration
     :param year: The year from the data to noise
+    :param verbose: Log with verbosity if True. Default is False.
     :return: A pd.DataFrame of noised W2 and 1099 data
     """
     year_filter = {"hdf": None, "parquet": None}
@@ -203,7 +238,9 @@ def generate_taxes_w2_and_1099(
         year_filter["hdf"] = [f"{FORMS.tax_w2_1099.date_column} == {year}."]
         year_filter["parquet"] = [(FORMS.tax_w2_1099.date_column, "==", year)]
         seed = seed * 10_000 + year
-    return _generate_form(FORMS.tax_w2_1099, source, seed, configuration, year_filter)
+    return _generate_form(
+        FORMS.tax_w2_1099, source, seed, configuration, year_filter, verbose
+    )
 
 
 def generate_women_infants_and_children(
@@ -211,6 +248,7 @@ def generate_women_infants_and_children(
     seed: int = 0,
     configuration: Union[Path, str, dict] = None,
     year: int = 2020,
+    verbose: bool = False,
 ) -> pd.DataFrame:
     """
     Generates noised Women Infants and Children (WIC) data from un-noised data.
@@ -219,6 +257,7 @@ def generate_women_infants_and_children(
     :param seed: An integer seed for randomness
     :param configuration: (optional) A path to a configuration YAML file or a dictionary to override the default configuration
     :param year: The year from the data to noise
+    :param verbose: Log with verbosity if True. Default is False.
     :return: A pd.DataFrame of noised WIC data
     """
     year_filter = {"hdf": None, "parquet": None}
@@ -226,7 +265,7 @@ def generate_women_infants_and_children(
         year_filter["hdf"] = [f"{FORMS.wic.date_column} == {year}."]
         year_filter["parquet"] = [(FORMS.wic.date_column, "==", year)]
         seed = seed * 10_000 + year
-    return _generate_form(FORMS.wic, source, seed, configuration, year_filter)
+    return _generate_form(FORMS.wic, source, seed, configuration, year_filter, verbose)
 
 
 def generate_social_security(
@@ -234,6 +273,7 @@ def generate_social_security(
     seed: int = 0,
     configuration: Union[Path, str, dict] = None,
     year: int = 2020,
+    verbose: bool = False,
 ) -> pd.DataFrame:
     """
     Generates noised Social Security (SSA) data from un-noised data.
@@ -242,6 +282,7 @@ def generate_social_security(
     :param seed: An integer seed for randomness
     :param configuration: (optional) A path to a configuration YAML file or a dictionary to override the default configuration
     :param year: The year up to which to noise from the data
+    :param verbose: Log with verbosity if True. Default is False.
     :return: A pd.DataFrame of noised SSA data
     """
     year_filter = {"hdf": None, "parquet": None}
@@ -251,4 +292,4 @@ def generate_social_security(
             (FORMS.ssa.date_column, "<=", pd.Timestamp(f"{year}-12-31"))
         ]
         seed = seed * 10_000 + year
-    return _generate_form(FORMS.ssa, source, seed, configuration, year_filter)
+    return _generate_form(FORMS.ssa, source, seed, configuration, year_filter, verbose)